+++ Anarchy/Emoji/Emoji/EmojisUpdatedEventArgs.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+
+namespace Discord.Gateway
+{
+    public class EmojisUpdatedEventArgs
+    {
+        public long GuildId { get; private set; }
+        public List<Emoji> Emojis { get; private set; }
+
+        internal EmojisUpdatedEventArgs(EmojiContainer emojis)
+        {
+            GuildId = emojis.GuildId;
+            Emojis = emojis.Emojis;
+        }
+    }
+}