+++ Anarchy/Emoji/Emoji/PartialEmoji.cs
@@ -2,43 +2,44 @@
 
 namespace Discord
 {
-    public class Emoji : Controllable
+    public class PartialEmoji : Controllable
     {
-        public Emoji()
-        {
-            OnClientUpdated += (sender, e) => Creator.SetClient(Client);
-        }
-
         [JsonProperty("id")]
-        public long Id { get; private set; }
+        public long? Id { get; private set; }
 
         [JsonProperty("name")]
         public string Name { get; private set; }
-        
+
         [JsonProperty("animated")]
         public bool Animated { get; private set; }
-
-        [JsonProperty("user")]
-        public User Creator { get; private set; }
 
         public long GuildId { get; internal set; }
 
 
         public void Update()
         {
-            Name = Client.GetGuildEmoji(GuildId, Id).Name;
+            if (Id == null)
+                return;
+
+            Name = Client.GetGuildEmoji(GuildId, (long)Id).Name;
         }
 
 
         public void Modify(string name)
         {
-            Name = Client.ModifyGuildEmoji(GuildId, Id, name).Name;
+            if (Id == null)
+                return;
+
+            Name = Client.ModifyGuildEmoji(GuildId, (long)Id, name).Name;
         }
 
 
         public void Delete()
         {
-            Client.DeleteGuildEmoji(GuildId, Id);
+            if (Id == null)
+                return;
+
+            Client.DeleteGuildEmoji(GuildId, (long)Id);
         }