+++ Anarchy/HTTP/DiscordHttpError.cs
@@ -5,7 +5,7 @@ namespace Discord
     public class DiscordHttpError
     {
         [JsonProperty("code")]
-        public int Code { get; private set; }
+        public DiscordError Code { get; private set; }
 
         [JsonProperty("message")]
         public string Message { get; private set; }