+++ Anarchy/Presence/Presence/PresenceUpdatedEventArgs.cs
@@ -1,0 +1,18 @@
+﻿namespace Discord.Gateway
+{
+    public class PresenceUpdatedEventArgs
+    {
+        public PresenceUpdate Presence { get; private set; }
+
+        public PresenceUpdatedEventArgs(PresenceUpdate presence)
+        {
+            Presence = presence;
+        }
+
+
+        public override string ToString()
+        {
+            return Presence.ToString();
+        }
+    }
+}