+++ Anarchy/User/User/UserPlatform.cs
@@ -1,0 +1,10 @@
+﻿namespace Discord
+{
+    public enum UserPlatform
+    {
+        None,
+        Desktop,
+        Mobile,
+        Web
+    }
+}